Refactor file imports for consistency and add .vite to .gitignore
