+++ .gitignore
@@ -2,7 +2,7 @@ Fagote.glb
 
 img/
 
-
+.vite
 
 # Logs
 logs
