+++ src/main.jsx
@@ -4,9 +4,9 @@ import { createBrowserRouter, RouterProvider } from 'react-router-dom'
 import App from './App'
 import './index.css'
 
-import Home from './pages/home/Home'
+import Home from './pages/Home'
 import About from './pages/About'
-import Hobbies from './pages/Hobbies'
+import Works from './pages/Works'
 import PageNotFound from './pages/PageNotFound'
 import ContentPage from './pages/ContentPage';
 import Contact from './pages/Contact'
@@ -18,7 +18,7 @@ const router = createBrowserRouter([
     children: [
       { index: true, element: <Home /> },
       { path: 'about', element: <About /> },
-      { path: 'hobbies', element: <Hobbies /> },
+      { path: 'works', element: <Works /> },
       { path: 'contact', element: <Contact /> },
       { path: '*', element: <PageNotFound /> },
       { path: ':name', element: <ContentPage /> }
