--- src/pages/Hobbies.jsx
@@ -1,10 +1,0 @@
-import Category from "../components/Category";
-
-export default function Hobbies() {
-    return ( 
-        <>
-        <Category/>
-        </>
-     );
-}
-
