+++ src/pages/Works.jsx
@@ -1,0 +1,10 @@
+import Category from "../components/Category";
+
+export default function Works() {
+    return (
+        <>
+            <Category />
+        </>
+    );
+}
+
