--- src/pages/home/style.css
@@ -1,3 +1,0 @@
-.who_am_I {
-    
-}
